+++ RC2.0/src/GenXSequence.hpp
@@ -44,7 +44,10 @@
 //=============================================================================
 
 
-#include "gpstkplatform.h"
+
+
+
+
 #include "PCodeConst.hpp"
 #include "mergePCodeWords.h"
 
@@ -139,14 +142,14 @@ namespace gpstk
           *  current value for lengthOfSequence, the function will wrap around
           *  to the beginning of sequence.
           */
-         uint32_t operator[] ( int i );
+         unsigned long operator[] ( int i );
          
          ///  Set the end of sequence for the current cycle.  
          void setLengthOfSequence( int i );
       
       private:
-         void addBitToSequence( uint32_t newBit );
-         uint32_t bits[MAX_WORD];
+         void addBitToSequence( unsigned long newBit );
+         unsigned long bits[MAX_WORD];
          int lengthOfSequence;
          int maxOfSequence;
          bool debugPrint;
