+++ RC2.0/src/X1Sequence.cpp
@@ -48,6 +48,7 @@
 
 
 
+
    // Project headers
 #include "Exception.hpp"
 #include "X1Sequence.hpp"
@@ -57,7 +58,7 @@ namespace gpstk
 {
       //   Static Variable Definition
    bool X1Sequence::isInit = false;
-   uint32_t* X1Sequence::X1Bits = 0;
+   unsigned long* X1Sequence::X1Bits = 0;
 
    X1Sequence::X1Sequence(  )
    {
@@ -85,7 +86,7 @@ namespace gpstk
          GPSTK_THROW(e);
       }
       
-      X1Bits = new uint32_t[NUM_6SEC_WORDS];
+      X1Bits = new unsigned long[NUM_6SEC_WORDS];
       if (X1Bits==0) 
       {
          gpstk::Exception e ("X1Sequence::allocateMemory() - allocation failed.");
@@ -99,8 +100,8 @@ namespace gpstk
          // Combination will be performed for four X1 epochs.
          // This will generate six seconds of X1 bits sequence
          // that will end on an even 32-bit boundary.
-      uint32_t X1Abits;
-      uint32_t X1Bbits;
+      unsigned long X1Abits;
+      unsigned long X1Bbits;
       X1Aepoch = 1;
       X1Acount = 0;
       X1Bepoch = 1;
